Break out previewer store/reducers.
