+++ ui/src/index.js
@@ -5,28 +5,7 @@ import { Provider } from 'react-redux';
 import Routes from './containers/Routes';
 import configureStore from './store/configureStore';
 
-const store = configureStore({
-    selectedChannel: 'desktop',
-    selectedLocale: null,
-    selectedType: null,
-    channels: {
-        desktop: {
-            name: 'Desktop',
-            localeIndexUrl: 'https://tiles.cdn.mozilla.net/desktop_tile_index_v3.json',
-            isFetching: false
-        },
-        prerelease: {
-            name: 'Prerelease',
-            localeIndexUrl: 'https://tiles.cdn.mozilla.net/desktop-prerelease_tile_index_v3.json',
-            isFetching: false
-        },
-        android: {
-            name: 'Android',
-            localeIndexUrl: 'https://tiles.cdn.mozilla.net/android_tile_index_v3.json',
-            isFetching: false
-        }
-    }
-});
+const store = configureStore({});
 
 React.render(
   <Provider store={store}>
